feat(station-service): throw an exception
- [지하철 역] 지하철 역이 지하철 노선에 등록된 경우 예외 처리하는 기능
구현

+++ src/main/java/subway/service/validation/NameDeletionValidation.java
@@ -1,18 +1,33 @@
 package subway.service.validation;
 
 import subway.repository.StationRepository;
+import subway.repository.TransitMapRepository;
 import subway.service.abstraction.validation.NameDeletionValidationInterface;
 import subway.view.output.station.StationExceptionView;
+
+import java.util.LinkedList;
+import java.util.List;
 
 public class NameDeletionValidation implements NameDeletionValidationInterface {
     @Override
     public boolean checkNameInTransitMap(String stationName) {
+        List<LinkedList<String>> transitMapsStationsNames
+                = TransitMapRepository.transitMapsStationNames();
 
+        for (LinkedList<String> transitMapStationNames : transitMapsStationsNames) {
+            if (transitMapStationNames.contains(stationName)) {
+                return true;
+            }
+        }
         return false;
     }
 
     @Override
     public boolean checkDeletionValidation(String stationName) {
+        if (checkNameInTransitMap(stationName)) {
+            StationExceptionView.printInvalidStationNameInTransitMapException();
+            return false;
+        }
         if (!StationRepository.deleteStation(stationName)) {
             StationExceptionView.printInvalidStationNameExistenceException();
             return false;
